+++ packages/terra-framework-docs/CHANGELOG.md
@@ -6,6 +6,9 @@
   * Updated `notification-dialog` and `tabs` examples.
   * Updated examples for `terra-slide-panel-manager`.
 
+* Added
+  * Added a `terra-slide-panel` example where the panel is open on page load and fullscreen.
+
 ## 1.19.0 - (April 12, 2023)
 
 * Added